chore: abstracting appointments and controllers
