+++ backend/src/app/controllers/AvailableController.js
@@ -1,15 +1,4 @@
-import {
-    startOfDay,
-    endOfDay,
-    setHours,
-    setMinutes,
-    setSeconds,
-    format,
-    isAfter,
-} from 'date-fns';
-import { Op } from 'sequelize';
-
-import Appointment from '../models/Appointment';
+import AvailableAppointmentService from '../services/AvailableAppointmentService';
 
 class AvailableController {
     async index(request, response) {
@@ -23,52 +12,9 @@ class AvailableController {
 
         const searchDate = Number(date);
 
-        const appointments = await Appointment.findAll({
-            where: {
-                provider_id: request.params.id,
-                canceled_at: null,
-                date: {
-                    [Op.between]: [
-                        startOfDay(searchDate),
-                        endOfDay(searchDate),
-                    ],
-                },
-            },
-        });
-
-        const schedule = [
-            '08:00',
-            '09:00',
-            '10:00',
-            '11:00',
-            '12:00',
-            '13:00',
-            '14:00',
-            '15:00',
-            '16:00',
-            '17:00',
-            '18:00',
-            '19:00',
-            '20:00',
-        ];
-
-        const available = schedule.map(time => {
-            const [hour, minute] = time.split(':');
-
-            const value = setSeconds(
-                setMinutes(setHours(searchDate, hour), minute),
-                0
-            );
-
-            return {
-                time,
-                value: format(value, "yyyy-MM-dd'T'HH:mm:ssxxx"),
-                available:
-                    isAfter(value, new Date()) &&
-                    !appointments.find(
-                        appoint => format(appoint.date, 'HH:mm') === time
-                    ),
-            };
+        const available = await AvailableAppointmentService.run({
+            provider_id: request.params.id,
+            date: searchDate,
         });
 
         return response.json(available);
